requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.4
django-crispy-forms==1.9.0
django-filter==2.2.0
django-modeltranslation==0.14.4
gunicorn==20.0.4
Pillow==7.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-gettext==4.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
